remove unnecessary portal
